requirements.txt
# Azure SKDs
azure-identity>=1.13.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=21.0.0
azure-mgmt-storage>=21.0.0

# Core dependencies
typer>=0.9.0
pyyaml>=6.0
